# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  /** Price in kobo (smallest currency unit) */
  priceKobo: number;
  /** @nullable */
  imageUrl: string | null;
  stock: number;
  category: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @minimum 1 */
  priceKobo: number;
  imageUrl?: string;
  /** @minimum 0 */
  stock: number;
  /** @minLength 1 */
  category: string;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 1 */
  priceKobo?: number;
  imageUrl?: string;
  /** @minimum 0 */
  stock?: number;
  category?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerEmail: string;
  customerName: string;
  status: OrderStatus;
  totalKobo: number;
  paystackReference: string;
  /** @nullable */
  paystackAccessCode?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPriceKobo: number;
  createdAt: string;
}

export type OrderWithItemsStatus = typeof OrderWithItemsStatus[keyof typeof OrderWithItemsStatus];


export const OrderWithItemsStatus = {
  pending: 'pending',
  paid: 'paid',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderWithItems {
  id: number;
  customerEmail: string;
  customerName: string;
  status: OrderWithItemsStatus;
  totalKobo: number;
  paystackReference: string;
  /** @nullable */
  paystackAccessCode?: string | null;
  createdAt: string;
  updatedAt: string;
  items: OrderItem[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface CartItem {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface CheckoutInput {
  customerEmail: string;
  /** @minLength 1 */
  customerName: string;
  /** @minItems 1 */
  items: CartItem[];
  callbackUrl: string;
}

export interface CheckoutResponse {
  orderId: number;
  paystackUrl: string;
  reference: string;
}

export interface StoreSummary {
  totalProducts: number;
  totalOrders: number;
  totalRevenueKobo: number;
  paidOrders: number;
  pendingOrders: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
};

export type ListOrdersParams = {
email?: string;
status?: string;
};