# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: orders
    description: Order management
  - name: checkout
    description: Payment and checkout
  - name: store
    description: Store summary and analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Products ──────────────────────────────────────────────────────────────
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Orders ────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (optionally filter by customer email)
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /orders/reference/{reference}:
    get:
      operationId: getOrderByReference
      tags: [orders]
      summary: Get order by Paystack reference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Checkout / Paystack ───────────────────────────────────────────────────
  /checkout/initiate:
    post:
      operationId: initiateCheckout
      tags: [checkout]
      summary: Create an order and get a Paystack payment URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /checkout/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [checkout]
      summary: Verify a Paystack payment and update order status
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Payment not successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /checkout/webhook:
    post:
      operationId: paystackWebhook
      tags: [checkout]
      summary: Paystack webhook receiver
      responses:
        "200":
          description: Acknowledged

  # ── Store summary (wow endpoints) ─────────────────────────────────────────
  /store/summary:
    get:
      operationId: getStoreSummary
      tags: [store]
      summary: Store-level stats — total products, orders, revenue
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSummary"

  /store/categories:
    get:
      operationId: listCategories
      tags: [store]
      summary: List distinct product categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

# ── Schemas ──────────────────────────────────────────────────────────────────
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Product ──────────────────────────────────────────────────────────────
    Product:
      type: object
      required: [id, name, description, priceKobo, imageUrl, stock, category, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        priceKobo:
          type: integer
          description: Price in kobo (smallest currency unit)
        imageUrl:
          type: ["string", "null"]
        stock:
          type: integer
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProductInput:
      type: object
      required: [name, description, priceKobo, stock, category]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        priceKobo:
          type: integer
          minimum: 1
        imageUrl:
          type: string
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        priceKobo:
          type: integer
          minimum: 1
        imageUrl:
          type: string
        stock:
          type: integer
          minimum: 0
        category:
          type: string

    # ── Order ────────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, customerEmail, customerName, status, totalKobo, paystackReference, createdAt, updatedAt]
      properties:
        id:
          type: integer
        customerEmail:
          type: string
        customerName:
          type: string
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]
        totalKobo:
          type: integer
        paystackReference:
          type: string
        paystackAccessCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    OrderItem:
      type: object
      required: [id, orderId, productId, productName, quantity, unitPriceKobo, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPriceKobo:
          type: integer
        createdAt:
          type: string
          format: date-time

    OrderWithItems:
      type: object
      required: [id, customerEmail, customerName, status, totalKobo, paystackReference, createdAt, updatedAt, items]
      properties:
        id:
          type: integer
        customerEmail:
          type: string
        customerName:
          type: string
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]
        totalKobo:
          type: integer
        paystackReference:
          type: string
        paystackAccessCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, shipped, delivered, cancelled]

    # ── Checkout ─────────────────────────────────────────────────────────────
    CartItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1

    CheckoutInput:
      type: object
      required: [customerEmail, customerName, items, callbackUrl]
      properties:
        customerEmail:
          type: string
        customerName:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
          minItems: 1
        callbackUrl:
          type: string

    CheckoutResponse:
      type: object
      required: [orderId, paystackUrl, reference]
      properties:
        orderId:
          type: integer
        paystackUrl:
          type: string
        reference:
          type: string

    # ── Store summary ─────────────────────────────────────────────────────────
    StoreSummary:
      type: object
      required: [totalProducts, totalOrders, totalRevenueKobo, paidOrders, pendingOrders]
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalRevenueKobo:
          type: integer
        paidOrders:
          type: integer
        pendingOrders:
          type: integer