# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface CheckoutInput {
  customerEmail: string;
  /** @minLength 1 */
  customerName: string;
  /** @minItems 1 */
  items: CartItem[];
  callbackUrl: string;
}