# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerEmail: string;
  customerName: string;
  status: OrderStatus;
  totalKobo: number;
  paystackReference: string;
  /** @nullable */
  paystackAccessCode?: string | null;
  createdAt: Date;
  updatedAt: Date;
}