# lib/api-zod/src/generated/types/orderWithItems.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderWithItemsStatus } from './orderWithItemsStatus';

export interface OrderWithItems {
  id: number;
  customerEmail: string;
  customerName: string;
  status: OrderWithItemsStatus;
  totalKobo: number;
  paystackReference: string;
  /** @nullable */
  paystackAccessCode?: string | null;
  createdAt: Date;
  updatedAt: Date;
  items: OrderItem[];
}