# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  /** Price in kobo (smallest currency unit) */
  priceKobo: number;
  /** @nullable */
  imageUrl: string | null;
  stock: number;
  category: string;
  createdAt: Date;
  updatedAt: Date;
}