# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @minimum 1 */
  priceKobo?: number;
  imageUrl?: string;
  /** @minimum 0 */
  stock?: number;
  category?: string;
}