# lib/api-zod/src/generated/types/storeSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoreSummary {
  totalProducts: number;
  totalOrders: number;
  totalRevenueKobo: number;
  paidOrders: number;
  pendingOrders: number;
}